how many times do I need to amend this?

+++ src/Main.java
@@ -3,7 +3,7 @@ public class Main {
         BmiService service = new BmiService();
         double weight = 98;
         double height = 1.87; // должно получиться 500
-        double bmi = service.calculate(height, weight);
+        int bmi = service.calculate(height, weight);
         System.out.println(bmi);
     }
 }
